backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.0
pydantic-settings>=2.0

# Database
sqlalchemy>=2.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0  # 同步数据库驱动（用于 OSS 配置读取）
alembic>=1.13.0
greenlet>=3.0.0

# Redis
redis>=5.0

# Storage
oss2>=2.18.0

# LLM
litellm>=1.23.0
anthropic>=0.18.0

# Agent Framework
langgraph>=0.0.26

# Task Queue & Scheduler
celery[redis]>=5.3.0
flower>=2.0.0  # Celery 监控面板（可选）
kombu>=5.3.0  # Celery 消息传递库

# Workflow Engine
temporalio>=1.4.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1  # 锁定版本，5.x 与 passlib 不兼容

# HTTP Client
httpx>=0.26.0

# Email
aioimaplib>=1.0.0
aiosmtplib>=3.0.0

# SSE
sse-starlette>=1.6.0

# PDF
reportlab>=4.0.0

# Feishu SDK
lark-oapi>=1.3.0

# Utils
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
